tic-tac-toe-game/tic-tac-toe-game/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
openai==0.27.0
python-dotenv==0.19.2
gunicorn==20.1.0